+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 2.0.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Supports additionalSchemata in field Add form
+  [ebrehault]
 
 
 2.0.5 (2015-06-05)